boiling_point_ingestion/requirements.txt
pandas
sqlalchemy
pgvector
jinja2
python-dotenv
transformers
torch